Use ApiError in ExactTargetService
